+++ src/main/java/me/myogoo/extendedterminal/init/wt/WTInits.java
@@ -1,23 +1,24 @@
 package me.myogoo.extendedterminal.init.wt;
 
-import appeng.api.features.GridLinkables;
-import appeng.items.tools.powered.WirelessTerminalItem;
-import de.mari_023.ae2wtlib.wut.WUTHandler;
 import me.myogoo.extendedterminal.me.host.ETWTHost;
+import me.myogoo.extendedterminal.menu.ETMenuType;
 import me.myogoo.extendedterminal.menu.extendedterminal.wt.ETWTMenu;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.RegisterEvent;
+import me.myogoo.myotus.api.wt.AddTerminalEvent;
 
-public class WTInits {
-    public static void init(RegisterEvent event) {
-        if (event.getRegistryKey().equals(ForgeRegistries.Keys.ITEMS)) {
-            WUTHandler
-                    .addTerminal("wireless_et_terminal",
-                            WTItems.WIRELESS_ET_TERMINAL.asItem()::tryOpen,
-                            ETWTHost::new,
-                            ETWTMenu.TYPE,
-                            WTItems.WIRELESS_ET_TERMINAL.asItem());
-        }
+public final class WTInits {
+    private WTInits() {
+    }
+
+    public static synchronized void registerTerminal() {
+        AddTerminalEvent.register(event -> {
+            var terminalName = ETMenuType.ET_TERMINAL.getWTIdAsString();
+            event.addTerminal(
+                    terminalName,
+                    ETWTHost::new,
+                    ETWTMenu.TYPE,
+                    WTItems.WIRELESS_ET_TERMINAL.asItem(),
+                    terminalName,
+                    "item.extendedterminal.wireless_et_terminal");
+        });
     }
 }