+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # CHANGELOG
 
+## Version Update
+- Require `Myotus` version to 15.0.7
+
+## Features
+- Extended Wireless Terminal support `Quantum Bridge Card`
+
 ## Bug Fixes
-- fix crash when open extended crafting guideme page  
+- fix JEI missingSlot miss match in shapeless crafting recipe
+- Fixed AE2WTLib Wireless Universal Terminal recipes to preserve Myotus upgrade storage data